Updates to console.
New: enriched the console with more markup, which gives richer presentation of the compiler output, i.e. colors (perhaps links or rollovers in the future).
Fix: improved the way Console.setStatus works, it was actually buggy before.
Fix: this was actually done in the previous changeset, printing of the url when errors are reported (unfortunately the line number is still incorrect)