Polishing rough edges. Setting up a spot for the documentation.
Fix: fixed the logInput function to respect the "enabled" (even though that is not so relevant) setting as well as update the console after registering the log message (which is important)
Fix: fixed the PolyMLext.onReady function (the PolyMLext.ready var was set in the wrong place). Though now the name "PolyMLext.ready" is not so appropriate anymore.
Fix: deploy script was trying to delete a wrong poly heap
Chg: removed the firstLaunch settings that was used for displayin the settings page after the first launch. That is not done anymore.
Chg: removed the tabs in the console, at least for now
Chg: renamed "about" to "settings" in all relevant places
New: added the minimize and off buttons in the console
New: started creating docs in the polymlext/chrome/content/docs